i fixed the error of the functuin of the task number 100

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -6,7 +6,7 @@
 
 int get_endianness(void)
 {
-	size_t temp = 1;
+	unsigned long int temp = 1;
 
 	return (*(char *)&temp);
 }
